requirements.txt
django
djangorestframework
djangorestframework-simplejwt
dj-database-url
gunicorn
psycopg2-binary
python-dotenv
